requirements.txt
mkdocs>=1.1
mkdocs-ringcentral>=0.1.20
markdown_fenced_code_tabs>=1.0.3
mkdocs-markdown-filter>=0.1.1
mkdocs-pymdownx-material-extras==1.0b11
mdx_include>=0.4.1
